src/requirements.txt
google-cloud-firestore==2.10.0
slack-bolt==1.16.4
slack-sdk==3.20.2
uvicorn==0.21.1
fastapi==0.95.0
aiohttp==3.8.4
google-api-python-client==2.82.0
